move readGrant-enpoints from contextRouting to its own routing
