+++ src/main/kotlin/no/bekk/routes/ReadGrantRouting.kt
@@ -1,0 +1,95 @@
+package no.bekk.routes
+
+import io.ktor.http.HttpStatusCode
+import io.ktor.http.parameters
+import io.ktor.server.plugins.BadRequestException
+import io.ktor.server.request.receive
+import io.ktor.server.request.receiveText
+import io.ktor.server.response.respond
+import io.ktor.server.routing.Route
+import io.ktor.server.routing.get
+import io.ktor.server.routing.patch
+import io.ktor.server.routing.post
+import io.ktor.server.routing.route
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.Json
+import no.bekk.authentication.AuthService
+import no.bekk.database.DatabaseReadGrantRequest
+import no.bekk.database.ReadGrantRepository
+import no.bekk.exception.ConflictException
+import no.bekk.plugins.ErrorHandlers
+import org.slf4j.LoggerFactory
+
+fun Route.readGrantRouting(authService: AuthService, readGrantRepository: ReadGrantRepository) {
+    val logger = LoggerFactory.getLogger("no.bekk.routes.AnswerRouting")
+
+
+    route("/readGrants") {
+
+        get {
+            logger.info("Received get /readGrants for user with id: ${call.parameters["userId"]}")
+            val userId = call.request.queryParameters["userId"] ?: throw BadRequestException("Missing userId parameter")
+
+            val usersReadGrants = readGrantRepository.getReadGrantsByUserId(userId)
+            call.respond(HttpStatusCode.OK, Json.encodeToString(usersReadGrants))
+            return@get
+        }
+
+
+        route("/{contextId}") {
+            get {
+                logger.info("Received GET /readGrants/contextId with id: ${call.parameters["contextId"]}")
+                val contextId = call.parameters["contextId"] ?: throw BadRequestException("Missing contextId")
+
+                if (!authService.hasWriteContextAccess(call, contextId) && !authService.hasReadContextAccess(call, contextId)) {
+                    call.respond(HttpStatusCode.Forbidden)
+                    return@get
+                }
+                val context = readGrantRepository.getReadGrantsByContext(contextId)
+                call.respond(HttpStatusCode.OK, Json.encodeToString(context))
+                return@get
+            }
+            post {
+                try {
+                    val contextId = call.parameters["contextId"] ?: throw BadRequestException("Missing contextId")
+
+                    val grantReadAccessRequestJson = call.receiveText()
+                    logger.info("Received POST /readGrants/contextId request with body: $grantReadAccessRequestJson")
+                    val grantReadAccessRequest = Json.decodeFromString<DatabaseReadGrantRequest>(grantReadAccessRequestJson)
+
+                    if (!authService.hasWriteContextAccess(call, contextId)) {
+                        call.respond(HttpStatusCode.Forbidden)
+                        return@post
+                    }
+
+                    val readGrant = readGrantRepository.insertReadGrantOnContext(contextId, grantReadAccessRequest)
+
+                    call.respond(HttpStatusCode.Created, Json.encodeToString(readGrant))
+                    return@post
+                } catch (e: ConflictException) {
+                    ErrorHandlers.handleConflictException(call, e)
+                } catch (e: Exception) {
+                    logger.error("Unexpected error when processing post /contexts/contextId/grantReadAccess", e)
+                    call.respond(HttpStatusCode.InternalServerError, "An unexpected error occurred.")
+                }
+            }
+
+            patch("{readGrantId}/expiry") {
+                logger.info("Received PATCH /readGrants/{contextId}/{readGrantId} with readGrantId: ${call.parameters["readGrantId"]}")
+                val contextId = call.parameters["contextId"] ?: throw BadRequestException("Missing readGrantId")
+                val readGrantId = call.parameters["readGrantId"] ?: throw BadRequestException("Missing readGrantId")
+
+                if (!authService.hasWriteContextAccess(call, contextId)) {
+                    call.respond(HttpStatusCode.Forbidden)
+                    return@patch
+                }
+
+                readGrantRepository.revokeReadGrantOnContext(readGrantId, contextId)
+                call.respond(HttpStatusCode.OK)
+                return@patch
+            }
+
+        }
+
+    }
+}
